fix: align AxonInfo/PrometheusInfo with subtensor, parallelize fetch calls

+++ src/types/prometheus.rs
@@ -1,23 +1,24 @@
 use serde::{Deserialize, Serialize};
 
 /// Prometheus information for a neuron
+/// Matches the subtensor pallet PrometheusInfo struct exactly
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct PrometheusInfo {
-    /// Block number when prometheus info was set
+    /// Prometheus serving block
     pub block: u64,
-    /// Version of prometheus info
-    pub version: u64,
-    /// IP address as string
+    /// Prometheus version
+    pub version: u32,
+    /// IP address as string (converted from u128)
     pub ip: String,
     /// Port number
     pub port: u16,
-    /// IP type (4 for IPv4, 6 for IPv6)
+    /// IP type: 4 for IPv4, 6 for IPv6
     pub ip_type: u8,
 }
 
 impl PrometheusInfo {
     /// Create PrometheusInfo from chain data - all fields required
-    pub fn from_chain_data(block: u64, version: u64, ip: String, port: u16, ip_type: u8) -> Self {
+    pub fn from_chain_data(block: u64, version: u32, ip: String, port: u16, ip_type: u8) -> Self {
         Self {
             block,
             version,
